requirements.txt
streamlit==1.35.0
folium==0.16.0
streamlit-folium==0.17.0
geopy==2.4.1
openrouteservice==2.3.8
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
polyline
